:recycle: (pc-12879) move PhoneFlip to correct folder
