+++ src/ui/svg/__tests__/PhoneFlip.test.tsx
@@ -1,11 +1,11 @@
 import * as React from 'react'
 
 import { render } from 'tests/utils'
-import { PhoneFlip } from 'ui/svg/PhoneFlip'
+import { PhoneFlip } from 'ui/svg/icons/PhoneFlip'
 
 jest.unmock('react-native-svg')
 
-describe('PhoneError', () => {
+describe('PhoneFlip', () => {
   it('should render component correctly', () => {
     const PhoneFlipIcon = render(<PhoneFlip />)
     expect(PhoneFlipIcon).toMatchSnapshot()
