+++ src/ui/svg/icons/PhoneFlip.tsx
